Move default network controller functions
